+++ changelog.md
@@ -8,6 +8,7 @@ Additions
 - Added reworked `ComponentProvider` and `ComponentContainer` interfaces to the V3 API
     - Refactored a lot of internal code to use the new types
 - Added `ComponentV3` and `ComponentRegistryV3` interfaces for early V3 adopters
+- Added a factory builder to the `ComponentContainer` interface
 
 Fixes
 - Fixed item components not being copied by the smithing table